added two samples for input phone number
