+++ playground/react/src/components/CountrySelect.tsx
@@ -8,76 +8,18 @@ interface CountryOption {
 }
 
 interface CountrySelectProps {
-  /**
-   * A two-letter country code.
-   * Example: "US", "RU", etc.
-   */
-  value?: Country;
-
-  /**
-   * A function of `value: string`.
-   * Updates the `value` property.
-   */
-  onChange: (value: Country) => void;
-
-  /**
-   * `<select/>` options.
-   */
-  options: CountryOption[];
-
-  /**
-   * `readonly` attribute doesn't work on a `<select/>`.
-   * To work around that, if `readOnly: true` property is passed
-   * to this component, it behaves analogous to `disabled: true`.
-   */
+  unicodeFlags?: boolean;
   disabled?: boolean;
   readOnly?: boolean;
   className?: string;
-}
-
-interface CountrySelectWithIconProps extends CountrySelectProps {
-  // Country flag component.
+  value?: Country;
+  options: CountryOption[];
   iconComponent?: ElementType<{ country: Country; label: string }>;
-
-  // Select arrow component.
   arrowComponent?: ElementType;
-
-  // Set to `true` to render Unicode flag icons instead of SVG images.
-  unicodeFlags?: boolean;
+  onChange: (value: Country) => void;
 }
 
-export const CountrySelect: FC<CountrySelectProps> = ({ value, onChange, options, disabled, readOnly, ...rest }) => {
-  const onChange_ = useCallback(
-    (event: ChangeEvent<HTMLSelectElement>) => {
-      const value = event.target.value as Country;
-      onChange(value);
-    },
-    [onChange],
-  );
-
-  return (
-    <select
-      {...rest}
-      disabled={disabled || readOnly}
-      value={value}
-      onChange={onChange_}
-    >
-      {options
-        .filter(({ value, divider }) => !divider && value)
-        .map(({ value: countryCode, label }) => (
-          <option
-            key={countryCode}
-            value={countryCode}
-            disabled={countryCode === value}
-          >
-            {label}
-          </option>
-        ))}
-    </select>
-  );
-};
-
-export const CountrySelectWithIcon: FC<CountrySelectWithIconProps> = ({
+export const CountrySelectWithIcon: FC<CountrySelectProps> = ({
   value,
   options,
   iconComponent: Icon,
@@ -100,10 +42,10 @@ export const CountrySelectWithIcon: FC<CountrySelectWithIconProps> = ({
   );
 
   return (
-    <div className='cb-phone-input-field'>
+    <div className='phone-input-field-button'>
       <select
         {...rest}
-        className='PhoneInputCountrySelect cb-phone-input-field-selection'
+        className='phone-input-field-selection'
         disabled={disabled || readOnly}
         value={value}
         onChange={onChange_}
@@ -115,22 +57,19 @@ export const CountrySelectWithIcon: FC<CountrySelectWithIconProps> = ({
               key={countryCode}
               value={countryCode}
               disabled={countryCode === value}
-              className={`cb-phone-input-field-selection-item${countryCode === value ? ' cb-phone-input-field-selection-item-selected' : ''}`}
+              className={`phone-input-field-selection-item${countryCode === value ? ' phone-input-field-selection-item-selected' : ''}`}
             >
               {label} (+{getCountryCallingCode(countryCode) || ''})
             </option>
           ))}
       </select>
-      <div className='cb-phone-input-field-button'>
-        {Icon && (
-          <Icon
-            country={value ?? 'US'}
-            label={(selectedOption && selectedOption.label) ?? ''}
-          />
-        )}{' '}
-        +{getCountryCallingCode(value ?? 'US') || ''}
-        <Arrow />
-      </div>
+      {Icon && (
+        <Icon
+          country={value ?? 'US'}
+          label={(selectedOption && selectedOption.label) ?? ''}
+        />
+      )}
+      <Arrow />
     </div>
   );
 };
