+++ playground/react/src/components/temp.css
@@ -1,0 +1,86 @@
+.phone-input-field {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 0rem;
+    width: 350px;
+    border: 1px solid #d1d1d1;
+    border-radius: 0.5rem;
+    height: 32px;
+}
+
+.phone-input-field input {
+    flex: 4;
+    width: 100%;
+    border: none;
+    font-size: 13px;
+    height: 93%;
+    border-top-right-radius: 0.5rem;
+    border-bottom-right-radius: 0.5rem;
+    padding-left: 0.5rem;
+}
+
+.phone-input-field input:focus {
+    outline: none;
+}
+
+.phone-input-field-button {
+    display: flex;
+    flex: 1;
+    position: relative;
+    align-self: stretch;
+    justify-content: center;
+    align-items: center;
+    background-color: #f8f8f8;
+    border: none;
+    cursor: pointer;
+    gap: 0.3rem;
+    height: 32px;
+    padding-left: 0.35rem;
+}
+
+.phone-input-field-button img {
+    width: 1.5rem;
+    margin: auto;
+}
+
+.phone-input-field-selection {
+    position: absolute;
+    top: 0;
+    left: 0;
+    height: 100%;
+    width: 100%;
+    z-index: 1;
+    border: 0;
+    opacity: 0;
+    cursor: pointer;
+    background-color: #f8f8f8;
+}
+
+.phone-input-field-selection-item {
+    display: flex;
+    align-items: center;
+    gap: 1rem;
+    padding: 1rem;
+    font-size: 13px;
+    cursor: pointer;
+    background-color: #f8f8f8;
+}
+
+.phone-input-field-selection-item-left {
+    display: flex;
+    align-items: flex-start;
+}
+
+.phone-input-field-selection-item-right {
+    display: flex;
+    align-items: flex-end;
+}
+
+.phone-input-field-selection-item:hover {
+    background-color: #a7a7a7;
+}
+
+.phone-input-field-selection-item-selected {
+    background-color: #a7a7a7;
+}
